Delete unused ancestor code
