+++ internal/pkg/chain/traversal_test.go
@@ -6,9 +6,11 @@ import (
 	"testing"
 
 	"github.com/filecoin-project/go-address"
+	"github.com/filecoin-project/specs-actors/actors/abi"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/filecoin-project/go-filecoin/internal/pkg/block"
 	"github.com/filecoin-project/go-filecoin/internal/pkg/chain"
 	th "github.com/filecoin-project/go-filecoin/internal/pkg/testhelpers"
 	tf "github.com/filecoin-project/go-filecoin/internal/pkg/testhelpers/testflags"
@@ -73,4 +75,163 @@ func TestIterAncestors(t *testing.T) {
 
 		assert.Error(t, it.Next())
 	})
+}
+
+// Happy path
+func TestCollectTipSetsOfHeightAtLeast(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+
+	chainLen := 15
+	head := builder.AppendManyOn(chainLen, block.UndefTipSet)
+
+	stopHeight := abi.ChainEpoch(4)
+	iterator := chain.IterAncestors(ctx, builder, head)
+	tipsets, err := chain.CollectTipSetsOfHeightAtLeast(ctx, iterator, stopHeight)
+	assert.NoError(t, err)
+	latestHeight, err := tipsets[0].Height()
+	require.NoError(t, err)
+	assert.Equal(t, abi.ChainEpoch(14), latestHeight)
+	earliestHeight, err := tipsets[len(tipsets)-1].Height()
+	require.NoError(t, err)
+	assert.Equal(t, abi.ChainEpoch(4), earliestHeight)
+	assert.Equal(t, 11, len(tipsets))
+}
+
+// Height at least 0.
+func TestCollectTipSetsOfHeightAtLeastZero(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+
+	chainLen := 25
+	head := builder.AppendManyOn(chainLen, block.UndefTipSet)
+
+	stopHeight := abi.ChainEpoch(0)
+	iterator := chain.IterAncestors(ctx, builder, head)
+	tipsets, err := chain.CollectTipSetsOfHeightAtLeast(ctx, iterator, stopHeight)
+	assert.NoError(t, err)
+	latestHeight, err := tipsets[0].Height()
+	require.NoError(t, err)
+	assert.Equal(t, abi.ChainEpoch(24), latestHeight)
+	earliestHeight, err := tipsets[len(tipsets)-1].Height()
+	require.NoError(t, err)
+	assert.Equal(t, abi.ChainEpoch(0), earliestHeight)
+	assert.Equal(t, chainLen, len(tipsets))
+}
+
+// The starting epoch is a null block.
+func TestCollectTipSetsOfHeightAtLeastStartingEpochIsNull(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+	head := builder.NewGenesis()
+
+	// Add 30 tipsets to the head of the chainStore.
+	head = builder.AppendManyOn(30, head)
+
+	// Now add 10 null blocks and 1 tipset.
+	head = builder.BuildOneOn(head, func(b *chain.BlockBuilder) {
+		b.IncHeight(10)
+	})
+
+	// Now add 19 more tipsets.
+	head = builder.AppendManyOn(19, head)
+
+	stopHeight := abi.ChainEpoch(35)
+	iterator := chain.IterAncestors(ctx, builder, head)
+	tipsets, err := chain.CollectTipSetsOfHeightAtLeast(ctx, iterator, stopHeight)
+	assert.NoError(t, err)
+	latestHeight, err := tipsets[0].Height()
+	require.NoError(t, err)
+	assert.Equal(t, abi.ChainEpoch(60), latestHeight)
+	earliestHeight, err := tipsets[len(tipsets)-1].Height()
+	require.NoError(t, err)
+	assert.Equal(t, abi.ChainEpoch(41), earliestHeight)
+	assert.Equal(t, 20, len(tipsets))
+}
+
+func TestFindCommonAncestorSameChain(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+	head := builder.NewGenesis()
+	// Add 30 tipsets to the head of the chainStore.
+	head = builder.AppendManyOn(30, head)
+	headIterOne := chain.IterAncestors(ctx, builder, head)
+	headIterTwo := chain.IterAncestors(ctx, builder, head)
+	commonAncestor, err := chain.FindCommonAncestor(headIterOne, headIterTwo)
+	assert.NoError(t, err)
+	assert.Equal(t, head, commonAncestor)
+}
+
+func TestFindCommonAncestorFork(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+	head := builder.NewGenesis()
+
+	// Add 3 tipsets to the head of the chainStore.
+	commonHeadTip := builder.AppendManyOn(3, head)
+
+	// Grow the fork chain
+	lenFork := 10
+	forkHead := builder.AppendManyOn(lenFork, commonHeadTip)
+
+	// Grow the main chain
+	lenMainChain := 14
+	mainHead := builder.AppendManyOn(lenMainChain, commonHeadTip)
+
+	forkItr := chain.IterAncestors(ctx, builder, forkHead)
+	mainItr := chain.IterAncestors(ctx, builder, mainHead)
+	commonAncestor, err := chain.FindCommonAncestor(mainItr, forkItr)
+	assert.NoError(t, err)
+	assert.Equal(t, commonHeadTip, commonAncestor)
+}
+
+func TestFindCommonAncestorNoFork(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+	head := builder.NewGenesis()
+
+	// Add 30 tipsets to the head of the chainStore.
+	head = builder.AppendManyOn(30, head)
+	headIterOne := chain.IterAncestors(ctx, builder, head)
+
+	// Now add 19 more tipsets.
+	expectedAncestor := head
+	head = builder.AppendManyOn(19, head)
+	headIterTwo := chain.IterAncestors(ctx, builder, head)
+
+	commonAncestor, err := chain.FindCommonAncestor(headIterOne, headIterTwo)
+	assert.NoError(t, err)
+	assert.True(t, expectedAncestor.Equals(commonAncestor))
+}
+
+// This test exercises an edge case fork that our previous common ancestor
+// utility handled incorrectly.
+func TestFindCommonAncestorNullBlockFork(t *testing.T) {
+	tf.UnitTest(t)
+	ctx := context.Background()
+	builder := chain.NewBuilder(t, address.Undef)
+	head := builder.NewGenesis()
+
+	// Add 10 tipsets to the head of the chainStore.
+	commonHead := builder.AppendManyOn(10, head)
+
+	// From the common ancestor, add a block following a null block.
+	headAfterNull := builder.BuildOneOn(commonHead, func(b *chain.BlockBuilder) {
+		b.IncHeight(1)
+	})
+	afterNullItr := chain.IterAncestors(ctx, builder, headAfterNull)
+
+	// Add a block (with no null) on another fork.
+	headNoNull := builder.AppendOn(commonHead, 1)
+	noNullItr := chain.IterAncestors(ctx, builder, headNoNull)
+
+	commonAncestor, err := chain.FindCommonAncestor(afterNullItr, noNullItr)
+	assert.NoError(t, err)
+	assert.Equal(t, commonHead, commonAncestor)
 }
